chore: address Oak's suggestion
