+++ apps/launch/components/Login/LoginV2.tsx
@@ -1,16 +1,9 @@
-import { SwapOutlined } from '@ant-design/icons';
 import { GetAccountReturnType, watchAccount } from '@wagmi/core';
-import { isNumber } from 'lodash';
-import { useCallback, useEffect, useMemo } from 'react';
+import { useCallback, useEffect } from 'react';
 import styled from 'styled-components';
-import { useAccount, useAccountEffect, useConfig, useDisconnect, useSwitchChain } from 'wagmi';
+import { useAccountEffect, useConfig, useDisconnect } from 'wagmi';
 
-import { useScreen } from 'libs/ui-theme/src';
 import { isAddressProhibited } from 'libs/util-prohibited-data/src/index';
-
-import { useAppSelector } from 'store/index';
-
-import { YellowButton } from './YellowButton';
 
 const LoginContainer = styled.div`
   display: flex;
@@ -22,10 +15,6 @@ const LoginContainer = styled.div`
 export const LoginV2 = () => {
   const { disconnect } = useDisconnect();
   const config = useConfig();
-  const { isMobile } = useScreen();
-  const { chain: walletConnectedChain } = useAccount();
-  const { networkId } = useAppSelector((state) => state.network);
-  const { switchChainAsync, isPending } = useSwitchChain();
 
   const handleConnect = useCallback(
     ({ address }: Pick<GetAccountReturnType, 'address' | 'chainId'>) => {
@@ -54,45 +43,8 @@ export const LoginV2 = () => {
     return () => unwatch();
   }, [config, clearOnDisconnect, handleConnect]);
 
-  /**
-   * @returns {boolean} - true if the wallet is connected to wrong network
-   * (ie. chain ID from wallet is different from the chain ID selected in the dropdown)
-   */
-  const isConnectedToWrongNetwork = useMemo(() => {
-    if (!isNumber(walletConnectedChain?.id) || !isNumber(networkId)) return false;
-
-    return walletConnectedChain?.id !== networkId;
-  }, [walletConnectedChain, networkId]);
-
-  const onSwitchNetwork = useCallback(async () => {
-    if (!networkId) return;
-
-    try {
-      await switchChainAsync({ chainId: networkId });
-    } catch (error) {
-      console.error(error);
-    }
-  }, [networkId, switchChainAsync]);
-
-  useEffect(() => {
-    if (isConnectedToWrongNetwork) {
-      onSwitchNetwork();
-    }
-  }, [isConnectedToWrongNetwork, onSwitchNetwork]);
-
   return (
     <LoginContainer>
-      {isConnectedToWrongNetwork && (
-        <YellowButton
-          loading={isPending}
-          type="default"
-          onClick={onSwitchNetwork}
-          icon={<SwapOutlined />}
-        >
-          {!isMobile && 'Switch network'}
-        </YellowButton>
-      )}
-      &nbsp;&nbsp;
       <w3m-button balance="hide" />
     </LoginContainer>
   );
